+++ frontend/package.json
@@ -9,6 +9,8 @@
     "axios": "^0.24.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-icons": "^4.3.1",
+    "react-image-button": "^1.0.6",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
     "react-simple-chatbot": "^0.6.1",
