+++ frontend/src/App.css
@@ -29,7 +29,7 @@
   background-size: contain;
   background-repeat: no-repeat;
   background-attachment: local;
-  background-color: #282c34;
+  background-color: #00A0FE;
   max-width: 100%;
   width: 100%;
   height: auto;
@@ -53,4 +53,17 @@
   to {
     transform: rotate(360deg);
   }
+}
+
+.chatbtn {
+  border: none;
+  color: none;
+  background-color: rgba( 255, 255, 255, 0 );
+  padding: 15px 32px;
+  text-align: center;
+  text-decoration: none;
+  display: inline-block;
+  font-size: 16px;
+  margin: 4px 2px;
+  cursor: pointer;
 }
